move RankRange into Rank.scala file
