+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/calcite/Rank.scala
@@ -138,6 +138,36 @@ abstract class Rank(
 
 }
 
+sealed trait RankRange extends Serializable {
+  def toString(inputFieldNames: Seq[String]): String
+}
+
+/** [[ConstantRankRangeWithoutEnd]] is a RankRange which not specify RankEnd. */
+case class ConstantRankRangeWithoutEnd(rankStart: Long) extends RankRange {
+  override def toString(inputFieldNames: Seq[String]): String = this.toString
+
+  override def toString: String = s"rankStart=$rankStart"
+}
+
+/** rankStart and rankEnd are inclusive, rankStart always start from one. */
+case class ConstantRankRange(rankStart: Long, rankEnd: Long) extends RankRange {
+
+  override def toString(inputFieldNames: Seq[String]): String = this.toString
+
+  override def toString: String = s"rankStart=$rankStart, rankEnd=$rankEnd"
+}
+
+/** changing rank limit depends on input */
+case class VariableRankRange(rankEndIndex: Int) extends RankRange {
+  override def toString(inputFieldNames: Seq[String]): String = {
+    s"rankEnd=${inputFieldNames(rankEndIndex)}"
+  }
+
+  override def toString: String = {
+    s"rankEnd=$$$rankEndIndex"
+  }
+}
+
 object Rank {
   def sortFieldsToString(collationSort: RelCollation): String = {
     val fieldCollations = collationSort.getFieldCollations
