+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/plan/nodes/logical/FlinkLogicalRank.scala
@@ -18,8 +18,7 @@
 package org.apache.flink.table.plan.nodes.logical
 
 import org.apache.flink.table.plan.nodes.FlinkConventions
-import org.apache.flink.table.plan.nodes.calcite.{LogicalRank, Rank}
-import org.apache.flink.table.plan.util.RankRange
+import org.apache.flink.table.plan.nodes.calcite.{LogicalRank, Rank, RankRange}
 
 import org.apache.calcite.plan._
 import org.apache.calcite.rel.`type`.RelDataType
